// src/sdks/systemV2ApiDocs.ts
export interface FactorGasReq {
  createBy?: number;

  createTime?: Date;

  factorId?: number;

  factorUnitM?: string;

  factorUnitZ?: string;

  factorValue?: string;

  gas?: string;

  gasType?: string;

  id?: number;

  updateBy?: number;

  updateTime?: Date;
}

export interface FactorGasRes {
  createBy?: number;

  createTime?: Date;

  factorId?: number;

  factorUnitM?: string;

  factorUnitZ?: string;

  factorValue?: string;

  gas?: string;

  gasType?: string;

  id?: number;

  updateBy?: number;

  updateTime?: Date;
  valid?: boolean;
}

export interface FactorReq {
  createBy?: number;
  createTime?: Date;

  description?: string;

  firstClassify?: string;

  gasList?: FactorGasReq[];

  id?: number;

  institution?: string;

  name?: string;

  secondClassify?: string;

  source?: string;

  sourceLanguage?: string;

  sourceLanguageName?: string;

  sourceLevel?: string;
  updateBy?: number;
  updateTime?: Date;

  url?: string;

  year?: string;
  languageSourceList?: {
    langType: number;
    sourceType: any;
    sourceValue: any;
    id: any;
  }[];
}